Rename route '/criteria' to '/criteria_stats'
Rename the route '/criteria', which shows criteria statistics, to
'/criteria_stats'. This makes the route consistent with '/project_stats'.
It also means we can use '/criteria' to present
the criteria themselves (in a format we control and in any locale).

Technically this could be considered an API change.
However, we only serve this as HTML, not as JSON (yet), so anyone
processing this would have had to request the HTML & process it further.
In practices, we expect that few people call this, because it
simply gives statistics about *just* the criteria while ignoring projects.
It's typically more been useful for debugging or giving people a
quick summary about the criteria than anything else.

Note: After merging, you'll need to send the YAML translations to
translation.io, because this renames several keys.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>