Reformat translation YAML files
This commit is the final state after uploading to translation.io
a variety of changes, followed by a 'rake translation:sync'.

The result is that we've reformatted the YAML files containing
translations into the "standard" format after a sync.
This was done so that we can easily detect changes to the translation
(since the YAML is now back to its normalized form).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>