Create text for criteria common introduction
This creates text to introduce the criteria text.

It's important to introduce the criteria text, but we also
want to be kind to the translators. So I've organized the
text so that, as much as possible, we can have *exactly*
the same text introduce the complete set of criteria *or*
a specific badge level set of criteria. That way the translators
only have to translate one batch of text.

The text is subdivided into many sub-headings, so that people
can simply skim the headings if they just want to get an idea
of what the summary is about without reading the details.

This commit required tweaking the tests. We now allow more HTML
tags and attributes, because they were needed for this
lengthy introduction.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>