Update rubocop (0.52.1->0.55.0)
Updated rubocop (style checker).

One challenge in updating rubocop, our style checker, is that it can
take a lot of effort to comply with all the added rules, and often there
isn't enough return on the time to justify it.

The solution taken here is that if there are a huge number of
violations, but they don't seem to be a serious problem, we modify
the configuration file .rubocop.yml to ignore them.
Otherwise we won't be able to update in reasonable time.
Also, much later versions of rubocop might be able to automate
more changes.

This also doesn't try to update to the latest version in a single step.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>