Refer to the *website* for the criteria
This changes many documents and the website front page
so that users are directed to /criteria to see the
actual criteria, instead of the static markdown files.

This does *NOT* change all the specialized README links to
specific criteria; the plan is to do that later.
Let's get the main links updated first :-).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>