Tweak /criteria_stats view (criteria link at bottom)
Tweak the /criteria_stats view by moving the /criteria
link to the bottom of the page. This makes more sense, because
now we have a page title followed *immediately* by its
main contents (the table), and only after that do we link
to related information. This commit doesn't change anything
for translators, because we're simply moving where it is on
the page, though a previous commit *did* affect translators
(because we're now linking to ourselves instead of GitHub).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>