Add url= query to /projects
To simplify dashboard use, add a new url= query parameter
that *specifically* searches URLs (this matches
either the home page URL or the repository URL).
If we add package URLs, we would search them as well.

We don't need to add a new index, because both of these
columns are already indexed.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>