Enable admin CLI search of email and name (for GDPR)
Add the ability for the BadgeApp admin to more easily search on
user email and user name. This is primarily so that we can easily
support GDPR requests, though it can have other uses.

The name search is relatively slow; instead of being indexed,
we look for internal matches while downcasing. But this is a
rare event that cannot be requested by unauthorized users,
so it doesn't matter. It's only *relatively* slow anyway; the
search happens entirely within the database.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>