Add ability to control cache size
Add the ability to set the environment variable RAILS_CACHE_SIZE
to control the cache size in Mebibytes (default 128 Mebibytes;
the old default was 64 Mebibytes). Warning: the Rails ".megabytes"
actually means Mebibytes.

I tried merging this into a single initializer, but that caused
test failures. I believe the problem is that that initializing the
cache is especially order-dependent (which makes sense, because it
affects so many other things). I'm sure I could track that down,
but trying to eliminate the duplication of a single simple line
isn't worth the trouble (especially since it was already duplicated).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>