Add HTTP Header Permissions-Policy (#1469)
Use the new HTTP security header, "Permissions-Policy", to disable things
we don't need. It was formerly named "Feature-Policy" with a slightly
different syntax. For now we'll generate both headers, for users with
slightly older web browsers.

This is a security hardening header. With this header, an attacker
will have fewer client privileges even if the attacker finds a XSS or
similar attack that works in our application.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>