Cache badge on CDN in JSON (in addition to SVG)
We need to add rate limits to counter overeager spidering tools
*without* impacting legitimate users. Some users (e.g., dashboards)
want to provides lots of data about project status.
We already use our CDN (Fastly) to cache the badge SVG.

This commit *also* caches the badge JSON data on our CDN.
That will let us support many more requests for badge JSON data.

The actual code change is tiny, but this adds many comments.
I had to do quite a bit a research about caches and CORS to
gain confidence that this wouldn't have some subtle problem in practice.

In the process, I discovered that our support for the
HTTP header 'Accept:' looks nice in theory, but in practice
creates complications. So I documented that we want people to use
different URLs with different suffixes to request different formats.
This is easy to understand and the app has always supported this.
We can remove support for 'Accept:' after it's been formally deprecated
for a while.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>