Optimize badge request - remove Origin from Vary
When requesting the badge itself (in JSON or SVG format),
optimize CDN results by *removing* Origin from the
HTTP Header "Vary". The badge data does not vary regardless
of who makes the request or what the origin is.

This optimizes the CDN because it will now have a single cached
value that it uses no matter what its
Origin value is (either as set by CORS or as not-set when
not using CORS). There will still be two badge formats
(SVG and JSON), and each will have up to two encodings
(compressed or not), but that's it. That increased sharing
increases the likelihood of being able to use a
previously-cached CDN value. Being able to use CDN values
in more cases helps us scale up greatly with few resources.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>