Switch gems from will_paginate pagy, fixes #1480 (#1482)
This commit changes how we paginate project lists when they're being
displayed. This switches from the will_paginate gem to the pagy gem
to speed view generation, use less memory, and use less code.

The will_paginate gem works, but the article
["Speeding Up Your Ruby on Rails App" by Daniel Lempesis](https://medium.com/@daniellempesis/speeding-up-your-ruby-on-rails-app-4c37ec71b126]
noted that while this was common, the
[pagy gem](https://github.com/ddnexus/pagy) was a significantly
better option. It's much faster and uses significantly less memory.
It's also just cleaner (in my opinion).
Although pagy works differently, switching looked pretty easy, and it was.

The tweak to the robots.txt file isn't strictly necessary, since that's
only needed if we use the JavaScript versions. But it would be easy to
switch to the JavaScript versions and forget to tweak the robots.txt
file, so this commit modifies the robots.txt file pre-emptively.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>