Gemfile: Break rails gem into selected dependencies (#1483)
Modify the Gemfile so we *only* load the Rails gems we use.
This reduces our memory use and attack surface.

Previously in our Gemfile we just used this construct to load Rails:

> gem 'rails', '5.2.4.4' # Our web framework

However, that brings in some Rails gems that we never use.

Bringing in unused gems can be a problem.
[Rails issue #36963, "Rails 6.0.0 performance regression because of ActionText::Engine hook"](https://github.com/rails/rails/issues/36963)
discusses a serious problem when using Rails 6 - ActionText
takes a lot more memory (so much that many people have reverted
an upgrade). We don't plan to use ActionText anyway;
it fails to support markdown.
ActionText builds on Trix, and we have requested markdown suport
([Trix issue #626](https://github.com/basecamp/trix/issues/626))
and someone has developed a pull request for some markdown support
([Trix pull request #737](https://github.com/basecamp/trix/pull/737)).
But those patches have yet to be accepted, and it's not clear this
functionality will ever get added. Without that functionality,
ActionText is not useful for many people (not just us).
So we'd have a lot of pain (wasted memory) with no benefit (we can't
use ActionText anyway).

The StackOverflow discussion
[Rails 5.2.4: How to reduce RAM use?](https://stackoverflow.com/questions/59340237/rails-5-2-4-how-to-reduce-ram-use)
noted several ways to reduce RAM use, including:

> Another thing you can do is to only load the rails modules you
> actually use. On your config/application.rb file you'll see a
> line like this require 'rails/all' that loads all rails features
> https://github.com/rails/rails/blob/master/railties/lib/rails/all.rb
> You can change that line to only import the features you want, like if
> you don't use action_cable or active_job you can just import the rest.

However, while we *could* modify `config/application.rb`, it seems
much cleaner to list in the Gemfile *only* the gems we use.
Then we *know* we never loaded anything else.

A minor negative to this approach is that when we update our Gemfile we
have to update all these gem versions in sync. Since they're released
as a unit, it's possible that different rails gems won't work properly
together if they use different versions. This commit adds
comments to the Gemfile to specifically counter that potential problem.
As a result, I think that risk is very low.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>