Add early version of assurance-case-verification.odg
This adds an early version of assurance-case-verification.odg.

Scott Ankrum is providing useful feedback to improve the
CII Best Practices assurances case. His goal is to create
a SACM version of the assurance case, using a SACM tool.
In the process, he's provided a lot of feedback on how to
improve our assurance case.

I haven't committed to switching from CAE notation to SACM notation.
Also, the tool they're using does not appear to be open source software
(a requirement on all tools needed to modify the BadgeApp)
and seems to be alpha-quality early stage software.
However, I definitely want to take advantage of his insights.

So I intend to work with him to modify our current assurance case
information, so that our assurance case in CAE notation can be
trivially converted to SACM and back.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>