assurance-case-verification.org: Updates
Add evidence. Change icon fills to more closely match:
https://www.adelard.com/asce/choosing-asce/cae.html
The icons shown are all white with only colored edges,
but the colors are hard to see; changing the icons
to be fully colored makes the color differences easier to see.
We continue to use their icons.
So this means we have

* oval (with blue fill) = claim/subclaim
* rounded rectangle (with green fill) = argument
* box (with purple fill) = evidence

Fills are intentionally light so they don't interfere with
reading the black text.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>