Add a new /badge_static/:value route (#1489)
* Add a new /badge_static/:value route

Add a new /badge_static/:value route; this serves *static* SVG
badge images given just "value". Value can be 0..99, passing,
silver, or gold. This will be cached by the CDN (Fastly).

This is the first step in implementing some countermeasures against
CDN race conditions. The plan is to use this to
serve /:locale/projects/:number badge entries. That way, if the
badge entry has just changed, but the CDN hasn't caught up yet, the
user who just change the project data will nevertheless
see the *current* status of the project.
We could also do this with the /projects display.

We still want projects, on their README, to use the
/projects/:number/badge URL when displaying badges, because that
will automatically adjust as the project's badge status changes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Refector badge_value into Project class

Refactor by moving badge_value into the Project class.
This will make it much easier to refer to it from a view.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>