Use /badge_static/ in /projects/NUM display
Switch to using /badge_static/:level in the /projects/NUM display.
This eliminates a race condition where the requestor is trying to
get the badge status image, but sends the request to the CDN
before the CDN has had a chance to update the caches local to
that user. This can happen, for example, when a user is editing.

This also reduces the impact if we have a problem updating the
CDN cache, e.g., because the CDN key has changed. If the CDN key
changes and we don't update our key, cache purges fail. With this in place,
users will be able to update a specific project and see the /projects
view with its *current* status, even though the CDN is still serving the
older data. Many people will be unsurprised if caches take time to update.
So while this isn't desired behavior, it would be unsurprising to many,
and the problem would sort itself out over time automatically.

There *is* a risk, namely, that people might copy the
/badge_static/:level URL instead of using the correct
/projects/:project_number/badge URL. I've included a comment
telling people *NOT* to use those URLs.

A future alternative might be to use the /badge_static URL if the
project was modified within 24 hours, else switch to the
/projects/:project_number/badge URL. I only just thought of this
while writing this commit, so I didn't do that :-).
Someone could do that in the future if desired.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>