Add Brakeman GitHub action
Create a .github/workflows/main.yml file and use it to
call Brakeman to do static analysis

Almost all of our CI pipeline runs on CircleCI.
However, GitHub actions provides a convenient way to run Brakeman,
a static analysis tool for Ruby on Rails applications.
There's no reason we can't use both, and I want to quicky
re-add Brakeman. We can move Brakeman to CircleCI later if
that's important.

This uses our credential, but note that we do NOT check in
the credential into version control (that would be bad!).
That does raise the question of whether not we trust the tool
and the GitHub action shim to call it.
GitHub themselves use Rails, so while I don't know devmasx
personally, I think GitHub's vetting is adequate for us to trust it.

Brakeman's license allows us to use it, but it is no longer OSS.
However, since we're not bringing the Brakeman code into our code,
and we don't *depend* on Brakeman to execute correctly for our code
to execute correctly, I think it's okay. I still intend to also get
Railroader integrated, but there's no reason we can't have both,
and I really want to get a static tool re-integrated. We temporarily
disabled Brakeman & I keep forgetting about this. Let's get a static
analysis tool integrated *now*, and later improve things.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>