Use project badge URL depending on updated_at
When showing a project's image src URL for its badge image (SVG),
use a different one depend on when it was last updated:

* If the project entry has changed relatively recently,
we give its /badge_static value.  That way, the user sees the
correct result even if the CDN hasn't completed distributing the
new value or a bad key prevents its update.
* If the project entry has NOT changed relatively recently,
we give the /projects/:id/badge value, so that humans who copy the
values without reading directions are more likely to see the URL that
we want them to use in READMEs. We also include a comment in the HTML
view telling people to use the /projects/:id/badge URL, all to encourage
humans to use the correct URL.

We're trying to strike a balance. Users who edit entries will always
see current results, but people who view this will eventually
see the preferred URL for badges.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>