Ensure project caches expire
The project displays now change over time (because the badge URL
changes if the project entry is NOT changed for a while).
Thus, we need to tell our caches about that.

The easy way is to give the caches an expiration time.
Technically that's not always needed, since once it changes to its
"old form" it will stay that way, but it can't hurt to
*ocassionally but rarely* recreate cached values.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>