Do not display email address in user display
Thanks to COVID-19, people often share screens with others.

This commit removes the screen display of email addresses on user display,
even to those authorized to see it. Instead, it just shows "Send email"
as the text. We continue to support a mailto: hyperlink.

No functionality is lost, this commit just reduces the risk of
unintentional disclosure via screen sharing.
Those who are authorized to see the email addresses can still
hover to see it, or click on the link to send an email, or
right-click to copy the email address.

We also still display the email address if a user is *editing* their
entry. But editing is a much rarer event, and we *have* to display
the email address then (so users can confirm it and/or edit it
to the correct value).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>