Switch top-level to mostly claims
Scott Ankrum (MITRE) has been redoing the assurance case
and strongly recommends rewording things as a sequence of
subclaims, and is recreating the assurance case in SACM.
It *is* easier to understand as subclaims.

This commit revises the top level as claims and sub-claims.
It follows many of his wordings, but with some changes.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>