Simplify code: Have a single CDN surrogate key (#1492)
Simplify the code by switching to a single CDN surrogate key.

We had two different CDN surrogate keys for each project,
one for the project badge SVG and one for the project badge as 1JSON.
But that doesn't make much sense. If the project data (as JSON) changes,
often the SVG will change too. So the separation creates extra work
and code for a relatively rare benefit.

Instead, use a single per-project key. That's much easier to understand.
When the project changes. then just purge that project's CDN key.
That's much clearer.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>