bundle update (update transitive dependencies)
With this update, `libyear-bundler` reports:

> System is 33.0 libyears behind

That is the total time, across all dependencies (direct and indirect),
between the version of libraries we use and the latest version available
of each of those libraries.

The command `rake report_code_statistics` reports:

> Number of gems (direct dependencies only) = 83
> Number of gems (including indirect dependencies) = 193

Thus, this version today has 2.05 average libmonths (33.0*12/193).

I think that's really good for an application that's been around
for over 6 years and has nearly 200 dependencies.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>