Put bad_password updater in model and test it
Put updater in model. That way it's easy to test... and now
test it.

One problem is that this creates a slow test. The easy solution
is to update to Rails 6, which provides a bulk insert.
So let's take the slow test for now, update to Rails 6, and then
modify the test to be faster.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>