script/update-database.sh: Added (#1518)
Add new script "update-database.sh". This automates the steps
to update the Heroku database. The Heroku site documents how to do it,
but there are a *lot* of steps and it's *critical* to do them *all* in the
correct order. Automating them increases the likelihood of doing them
correctly.

This was used to update our site from PostgreSQL 11 to PostgreSQL 12
(the latest version Heroku supports).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>