Switch test env to CircleCI cimg
CircleCI is now promoting the use of a new set of starting
images in the "cimg" space, designating the ones we used
as "legacy". The newer "cimg" images are built on Ubuntu,
and are supposed to be faster and more deterministic.
See <https://hub.docker.com/r/cimg/ruby>.

This commit:
- creates a new Dockerfile for a special image
  that builds on this newer "cimg" format
  `dockerfiles/2.7.2-browsers/Dockerfile` that
  adds what we need (e.g., cmake).
- Switches to using this newer "cimg" format on CircleCI
  for testing and building.

This was *incredibly* painful to achieve, and took *much*
longer than it should have taken.

CircleCI strongerly recommends using the browser-tools orb.
However, I never got that to work in spite of endless tweaking.
I ended up creating a special task to install a browser (Chrome).

The result was *very* slow because the cache wasn't caching anything.
This setup ends up putting Ruby gems in ~/.rubygems, which is
not what I would have expected. Once the cache was reconfigured,
it works as expected. As always, changes to Gemfile.lock require
rebuilding gems (and thus takes extra time), but many changes don't
change Gemfile.lock, so caching is very much worth doing.

I had to change how gem omniauth-rails_csrf_protection was installed,
so it could be properly cached, but that is worth it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>