Tweak criterion test, fixes #1509
Add this text to the criterion "test":

> The project MUST clearly show or document how to run the
> test suite(s) (e.g., via a continuous integration (CI)
> script or via documentation in files such as BUILD.md, README.md,
> or CONTRIBUTING.md).

This was in response to a comment by dcmiddle, who noted that:

> Often projects contain multiple languages and testing features composed
> over different components.  Continuous integration tests may then call
> scripts which run these kinds of tests and those scripts may not look
> like e.g. make check.

Technically this is a change to this criterion. However, the only way
to show that a project uses a test suite (and thus meets the original
criterion) is to show it or document it. So in reality this text was always
implied. This change merely turns an implied requirement into a
requirement that is more clearly stated.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>