Tweak release_notes_vulns, fixes #1508
One project interpreted this criterion to
"add a CVE to each vulnerability fixed in the release".

We don't *oppose* doing that, but that wasn't the intent.
Instead, the intent was that "if a vulnerability reported as a CVE has
been fixed [then] the relevant CVE should be mentioned in the
release notes."

This rewords the criterion text so that it's hopefully clearer.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>