Force update gem nokogiri to '1.11.0.rc4'
This counters CVE-2020-26247,
"XML::Schema input is now "untrusted" by default".
For more information, see:
https://github.com/sparklemotion/nokogiri/security/advisories/GHSA-vr8q-g5c7-m54m

This is debatable. We are almost certainly *not* vulnerable, because we
never use Nokogiri's XML Schema system. In addition, this recommended
fix uses a *release candidate* version of a library, instead of a final
release, and we generally avoid doing that. However, it's hard to be
*certain* that there's no exploit, and this release candidate
appears to be very solid for our purposes. I think there
are good arguments to *use* this nokogiri release candidate, and that there
are also good reasons *not* to use it. Sadly I must make *some* decision;
no decision is a decision. I've made a judgement call to
use the release candidate version with the security update.
I think the risk is smallest if we upgrade now for this specific case.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>