Initial attempt to use Rails 6.0.
Initial attempt to use Rails 6.0.

Note that `bundler update` fails because we *must* update
`rails-i18n`, but that's incompatible with `fastly-rails` which
cannot handle Rails 6:

~~~~
Bundler could not find compatible versions for gem "railties":
  In Gemfile:
    fastly-rails (= 0.8.0) was resolved to 0.8.0, which depends on
      railties (> 2, < 6)

    rails-i18n (= 6.0.0) was resolved to 6.0.0, which depends on
      railties (< 7, >= 6.0.0)
~~~~

This means we must replace `fastly-rails` with `fastly-ruby`,
but `fastly-ruby` has a different API, so it will be nontrivial.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>