Remove gem fastly-rails
The gem "fastly-rails" does not work on Rails 6, so we
must eliminate it to upgrade to Rails 6.

They recommend switching to the "fastly" gem (aka "fastly-ruby"),
but fastly-ruby is not designed to support multi-threading (WHAT?!).
The only workable solution I could find was to modify our code
to directly call the Fastly API instead.

This turned out to be more complicated than expected, and
frankly was a big pain. However, this is a blocker to upgrading
Rails, and the version of Rails we use is no longer supported,
so this needed to be done.

We need to test this on the *real* system to ensure that it really works.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>