Omit session cookie if empty
By default Rails *always* sets a session cookie, even if it's empty.
That interferes with many cache optimizations and is generally
unfriendly. I tried to set an after filter to disable this, but that
happens at the wrong time. There's an ugly hack that worked on
Rails 3 but doesn't any more.

So let's disable sending empty session cookies for a few especially
popular pages at least.  Some discussion here:
https://stackoverflow.com/questions/5435494/rails-3-disabling-session-cookies

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>