Do not set *unchanged* session cookies in some cases
Don't send a session cookie if it's unchanged
(including "there was no value in it) in several cases.

Rails sends session cookies by default. This disables many
cache optimizations, and sending unnecessary cookies isn't
good for performance and privacy anyway. It's harder to omit
them in general, because flash error messages are part of a session
and Rails doesn't provide an easy mechanism to omit them in general.
So this commit omits unchanged session cookies in common cases,
so that the cookie is less commonly seen in practice.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>