Support LICENSES directory, fixes #1544
Modify details of the license_location criteria to expressly support
having a LICENSES directory.

In addition, modify the automation code to look for this LICENSES
directory, and presume that license_location is probably met (confidence
level 3/5) if a LICENSES directory is present at the top level.
We don't give it higher confidence (which would FORCE it) because
the directory might merely contain a discussion about licenses instead
of actually stating the licenses. That said, we're more likely to
determine the correct answer if a LICENSES directory exists, and that
is the point.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>