Update gem nokogiri (1.11.0.rc4 -> 1.11.1)
Update the gem nokogiri. If you look at the diff you'll see something
odd; this commit *removes* nokogiri as a direct dependency.

The reason is that nokogiri <1.11 has a publicly known vulnerability.
This was fixed in release candidate 1.11.0.rc4. However, by default
bundler does *not* use release candidates when selecting transitive
dependencies. So we forced the use of the release candidate by
expressly depending on the known-fixed release candidate as a direct
dependency. Now that we have a "normal" fixed release, we don't need to
claim it's a direct dependency; bundler will automatically determine
that we depend on nokogiri and choose a valid version. Bundler always
chooses the "most recent" version that meets its criteria, and now
a fixed version meets those criteria.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>