Move from Rails 5.2.X -> Rails 6.0.X
BIG MILESTONE! This commit *finally* switches the application
from the Rails 5.2.X series to the Rails 6.0.X series.

We've wanted to do this for a while, but other gems (libraries)
had blocked our way. The big problems (that we've been working to
overcome) were:

* We had to eliminate and replace the gem fastly-rails.
  The gem "fastly-rails" does not work on Rails 6, so we
  had to eliminate it to upgrade to Rails 6.
  They recommend switching to the "fastly" gem (aka "fastly-ruby"),
  but fastly-ruby is not designed to support multi-threading (WHAT?!).
  The only workable solution I could find was to modify our code
  to directly call the Fastly API instead.
  For details, see commit 2c9c0940ce9b42a00e003e6c2e4ec6b1012d9e35.

* We had to eliminate the gem minitest-rails-capybara,
  which we depended on for system tests, and modify all our system
  tests to work with the now-standard Rails system test system.
  The gem minitest-rails-capybara support ends with minitest-rails 3.x,
  and thus we *cannot* use Rails 6 as long as we used it for tests.
  For details, see commit fdbbc1eddd9ce78a232e6117175b4f80badd6cd0.

There was no rocket science, just a non-trivial amount of tedious work.

The XKCD cartoon "Software Updates" <https://xkcd.com/2224/>
illustrates the problems we had upgrading Rails, where
"an update finally breaks a feature I'm unwilling to lose"
and thus updating stops. In our case, we were unwlling to lose
our Content Delivery Network (CDN) since it's critical for performance,
and we were unwilling to lose our system tests because they help us
maintain high quality. So we had to put in some effort so that we could
do our update. If your update is difficult, it can create a security
problem because people may find it impractical to update!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>