Update to Rails 6.1 (partially done)
This commit *begins* an update from Rails 6.0.X -> 6.1.X.
This requires a large number of simultaneous library updates
(some took a while to identify).

This does *not* complete the update.

The biggest problem is the many error reports of this form:

> Error:
> ProjectsControllerTest#test_should_fail_to_create_project_with_duplicate_repo:
> ActiveRecord::StatementInvalid: PG::UndefinedColumn:
> ERROR:  column users.email does not exist

It's true that there is no users.email column in the database,
but that's because it's a virtual column that is *supposed* to be
managed by the `attr_encrypted` gem.
Since it's not being handled, it *appears* that this gem does not
work with ActiveRecord 6.1. I went to check out its status, and they
are looking for new maintainers:
https://github.com/attr-encrypted/attr_encrypted/issues/379
All options are not the desired ones here.

In addition, there are at least two kinds of deprecation warnings
which will need to be addressed (probably many times):

* DEPRECATION WARNING: action_view.raise_on_missing_translations is deprecated and will be removed in Rails 6.2. Set i18n.raise_on_missing_translations instead. Note that this new setting also affects how missing translations are handled in controllers. (called from call at /home/dwheeler/best-practices-badge/config/initializers/canonical_trailing_slash.rb:30)

* DEPRECATION WARNING: Rendering actions with '.' in the name is deprecated: static_pages/error_404.html.erb (called from error_404 at /home/dwheeler/best-practices-badge/app/controllers/static_pages_controller.rb:46)

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>