Update gem web-console (3.7.0->4.1.0)
This updates gem web-console, an in-browser debugger.
Insert <% console %> or console into the source code to use it.
This gem has been transferred to be part of the "rails" project.

Note that we only include this gem during development,
*not* in production. This gem enables running arbitrary code,
which would be a security problem if it was included in production.
But we do not call it from production code, and we don't include
the gem in production mode, preventing a security issue.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>