Add migration+change tests for blind_index update
The update to the blind_index gem requires a number
of changes no matter what.

This commit switches the application to the newer
blind_index naming conventions and algorithm (argon2id),
requiring a migration to do it and changes to the
test fixture for user data.

This commit was more work than expected.

Calculating the bidx values doesn't work correctly
from the fixtures within the ERB (the calls *run* but don't
produce the right values, presumably because it's not fully set up
at that point). This was resolved by inserting the calculated values
into the fixtures, which also makes the tests more independent
of the system being tested.
One of the older test values was subtly wrong (.com vs. .org
email address), which created a mysterious hard-to-track failure
(we prevent future versions of this problem
via a new test that sanity-checks all test values).

The migration had to be tweaked several times to work with
real production data. We need to not change the user updated_at
by setting touch to false (the user-visible data isn't changing),
we need not validate (some entries may not be acceptable any more
but we still need to migrate them as they are), and we need to
skip cases where encrypted_email is nil (in those cases
we have no data to decrypt).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>