Update gem pagy (3.10.0->3.11.0), Swahili built in
Update to the gem pagy version 3.11.0, and *eliminate*
the special-case pagy code for the Swahili translation
because we contributed it to the upstream pagy project.

When we first added Swahili, we had to special-case the
pagy configuration because pagy didn't include a Swahili
translation. We then contributed the pagy translation to
the pagy project (our upstream). Pagy incorporated that
Swahili translation and released a new version.
Now (by upgrading) we can eliminate our special code, making *our*
code simpler *and* making it easier for others to
support Swahili if they use pagy.

This is exactly the sort of flow that makes open source software
projects work. By getting improvements upstream,
the upstream projects are better and all downstream projects
reap the benefits (such as having simpler code).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>