Split up criteria_discussion.content_html
The key "criteria_discussion.content_html" is extremely long,
and as a result it's extremely inconvenient to our
hard-working translators.

This commit splits up the text into a large number of
smaller segments to translate. This should be much easier
to deal with. This split also makes it much easier to
match sentences to create parallel texts (e.g., for OPUS).

The keys are given names based on their major subsections.
That way, the subsections could be more easily moved around
(if needed).

This change does mean that changing the content of the
criteria_discussion view at
`app/views/static_pages/criteria_discussion.html.erb`
is more likely to require some editing.
However, this change makes it easier for our poor human translators,
and that's a trade-off worth making.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>