project_stats: cache does not depend on data size
The /project_stats HTML previously changed when the
data size changed. That is no longer true, so don't
have the cache depend on this information.

This has two positive effects: (1) it eliminates an
unnecessary round-trip call to the database (especially important
if we *do* have the page cached), and (2) it lets the
cached data live longer (since we no longer need to
regenerate it).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>