Use JSON.fast_generate in project_stats
Switch to using JSON.fast_generate in the project_stats
entries when they generate JSON. The datasets we create
can't have cycles, but the default generator doesn't know that, and
it is a nontrivial effort to do safety checks for cycles.

This commit generates many JSON entries using fast_generate,
skipping the unnecessary safety check for cycles.

The project_stats endpoints often generate a lot of JSON,
so this is worth doing. A benchmark of nontrivial_projects showed
this saved an average of 173ms and caused 42K fewer allocations
(172K->130K) for just that endpoint.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>